requirements.txt
# APK Multi-Agent Analyzer
# Python依赖

# Agent框架
langchain>=0.1.0
langchain-anthropic>=0.1.0
crewai>=0.28.0

# Anthropic SDK
anthropic>=0.18.0

# APK分析
androguard==3.4.0a1

# 配置文件
pyyaml>=6.0

# 报告生成
jinja2>=3.1.0
markdown>=3.5.0

# 数据处理
python-dateutil>=2.8.0

# 工具
colorama>=0.4.6

# 可选: 代码高亮
Pygments>=2.17.0

# 可选: 图表
matplotlib>=3.8.0

# 可选: 恶意软件检测
quark-engine>=24.4.0